Updated to work with ButtonPose dataset.

+++ unitytoocclusion.py
@@ -7,27 +7,39 @@ from math import ceil
 import yaml
 import json
 import os
+import cv2
+import numpy as np
 from os.path import join
 
 def main():
-    datalocationpath = './datasets/ScrewPose/data/'
-    numberOfSamples = 10000
-    trainamount = 9000
+    datalocationpath = './datasets/ButtonPose/data/'
+    numberOfSamples = 20000
+    trainamount = 18000
     samplesPerCaptureFile = 150
-    nameToIdDict = {"M8x50": 1, "M8x25": 2, "M8x16": 3, "M6x30": 4}
-    Cam_intrinsic = [905., 0., 640., 0., 905., 360., 0.0, 0., 1.]
 
-    deletePreviousGT(datalocationpath)
+    #nameToIdDict = {"M8x50": 1, "M8x25": 2, "M8x16": 3, "M6x30": 4} # ScrewPose
+    nameToIdDict = {"2-slot": 1, "3-slot": 2, "mushroombutton": 3, "arrowbutton": 4, "redbutton": 5, "unknownbutton": 6} # ButtonPose
+    nameToMaskDict = {"2-slot": 42, "3-slot": 84, "mushroombutton": 126, "arrowbutton": 168, "redbutton": 210, "unknownbutton": 255}
+
+    #Cam_intrinsic = [905., 0., 640., 0., 905., 360., 0.0, 0., 1.] # Realsense parameters
+    Cam_intrinsic = [640.0, 0., 640.0, 0., 640.0, 360.0, 0., 0., 1.] # Azure paremeters
+
+    deletePreviousGT(datalocationpath, len(nameToIdDict))
+
+    occlusionDict = loadOcclusionData(join(datalocationpath, "occlusions.txt"))
     
-    createLabels(datalocationpath, samplesPerCaptureFile, numberOfSamples, Cam_intrinsic, nameToIdDict)
+    createLabels(datalocationpath, samplesPerCaptureFile, numberOfSamples, Cam_intrinsic, nameToIdDict, occlusionDict)
 
     createSplits(datalocationpath, trainamount, numberOfSamples)
 
-    #renameImages(datalocationpath, numberOfSamples)
+    renameImages(datalocationpath, numberOfSamples)
 
-    print("Done!")
+    if occlusionDict:
+        correctMasks(datalocationpath, occlusionDict, nameToMaskDict)
 
-def createLabels(datalocationpath, samplesPerCaptureFile, numberOfSamples, Cam_intrinsic, nameToIdDict):
+    print("All tasks finished.")
+
+def createLabels(datalocationpath, samplesPerCaptureFile, numberOfSamples, Cam_intrinsic, nameToIdDict, occlusionDict):
     print("Creating ground truth labels...")
 
     totalCaptures = ceil(numberOfSamples/samplesPerCaptureFile)
@@ -48,6 +60,7 @@ def createLabels(datalocationpath, samplesPerCaptureFile, numberOfSamples, Cam_i
                 jsonData = json.load(file)    
 
                 for i in range(0, len(jsonData['captures'])):
+                    print("Progress: " + str(j + 1) + "/" + str(numberOfSamples), end='\r')
 
                     for annotation in jsonData['captures'][i]['annotations']:
 
@@ -69,6 +82,15 @@ def createLabels(datalocationpath, samplesPerCaptureFile, numberOfSamples, Cam_i
                                 imageData[name].translation = [1000*translationdata['x'], -1000*translationdata['y'], 1000*translationdata['z']]
                                 imageData[name].rotation = convertQuaternionToMatrix(rotationdata['x'], rotationdata['y'], rotationdata['z'], rotationdata['w'])
 
+                    if j in occlusionDict:
+                        imageData["unknownbutton"].present = True
+                        imageData["unknownbutton"].rotation = imageData[occlusionDict[j]].rotation
+                        imageData["unknownbutton"].translation = imageData[occlusionDict[j]].translation
+                        imageData["unknownbutton"].bbox = imageData[occlusionDict[j]].bbox
+
+                        imageData[occlusionDict[j]].clear()
+
+
                     datalist = []
                     
                     for key in imageData.keys():
@@ -84,10 +106,13 @@ def createLabels(datalocationpath, samplesPerCaptureFile, numberOfSamples, Cam_i
     print("Creating info labels...")
     with open(datalocationpath + 'info.yml', 'w') as ymlfile:
         for i in range(0, numberOfSamples):
+            print("Progress: " + str(i + 1) + "/" + str(numberOfSamples), end='\r')
             data = {i:{'cam_K':Cam_intrinsic, 'depth_scale':1.0}}
             yaml.dump(data, ymlfile)
 
 def convertQuaternionToMatrix(qx, qy, qz, qw):
+    # Transforms a quaternion rotation to its equivalent matrix form.
+
     qw = -qw
     qy = -qy
     R11 = 1 - 2 * (qy * qy + qz * qz) 
@@ -105,6 +130,9 @@ def convertQuaternionToMatrix(qx, qy, qz, qw):
     return R
 
 def createSplits(datalocationpath, trainamount, numberOfSamples):
+    # Creates the files specifying what files to use for training and what
+    # files to use for testing the model.
+
     print('Creating test/train splits...')
     with open(datalocationpath + 'train.txt', 'w') as file:
         for i in range(0, trainamount):
@@ -130,31 +158,82 @@ def createSplits(datalocationpath, trainamount, numberOfSamples):
 
 def renameImages(datalocationpath, numberOfSamples):
     print('Renaming training images...')
-    for i in range(0, numberOfSamples):
-        if i<10:
-            os.rename(datalocationpath + "rgb/rgb_" + str(i + 2) + ".png", datalocationpath + 'rgb/000' + str(i) + '.png')
-            os.rename(datalocationpath + "merged_masks/segmentation_" + str(i + 2) + ".png", datalocationpath + 'merged_masks/000' + str(i) + '.png')
-        elif i<100:
-            os.rename(datalocationpath + "rgb/rgb_" + str(i + 2) + ".png", datalocationpath + 'rgb/00' + str(i) + '.png')
-            os.rename(datalocationpath + "merged_masks/segmentation_" + str(i + 2) + ".png", datalocationpath + 'merged_masks/00' + str(i) + '.png')
-        elif i<1000:
-            os.rename(datalocationpath + "rgb/rgb_" + str(i + 2) + ".png", datalocationpath + 'rgb/0' + str(i) + '.png')
-            os.rename(datalocationpath + "merged_masks/segmentation_" + str(i + 2) + ".png", datalocationpath + 'merged_masks/0' + str(i) + '.png')     
-        else:
+    if(os.path.exists(join(datalocationpath, "rgb/rgb_2.png"))):
+        for i in range(0, numberOfSamples):
+            print("Progress: " + str(i + 1) + "/" + str(numberOfSamples), end='\r')
             os.rename(datalocationpath + "rgb/rgb_" + str(i + 2) + ".png", datalocationpath + 'rgb/' + str(i) + '.png')
             os.rename(datalocationpath + "merged_masks/segmentation_" + str(i + 2) + ".png", datalocationpath + 'merged_masks/' + str(i) + '.png')
+    else:
+        print("Found already renamed training images.")
 
-def deletePreviousGT(dirpath):
-    os.remove(join(dirpath, "gt.yml"))
-    os.remove(join(dirpath, "info.yml"))
-    os.remove(join(dirpath, "test.txt"))
-    os.remove(join(dirpath, "train.txt"))
-    os.remove(join(dirpath, "valid_poses/1.txt"))
-    os.remove(join(dirpath, "valid_poses/2.txt"))
-    os.remove(join(dirpath, "valid_poses/3.txt"))
-    os.remove(join(dirpath, "valid_poses/4.txt"))
+def deletePreviousGT(dirpath, objectnumber):
+    # Eliminates previous versions of ground truths in the given path.
+
+    if(os.path.exists(join(join(dirpath, "gt.yml")))):
+        os.remove(join(dirpath, "gt.yml"))
+
+    if(os.path.exists(join(join(dirpath, "info.yml")))):  
+        os.remove(join(dirpath, "info.yml"))
+
+    if(os.path.exists(join(join(dirpath, "test.txt")))):
+        os.remove(join(dirpath, "test.txt"))
+
+    if(os.path.exists(join(join(dirpath, "train.txt")))):
+        os.remove(join(dirpath, "train.txt"))
+
+    if(os.path.exists(join(join(dirpath, "valid_poses")))):
+        valid_pose_dir = join(dirpath, "valid_poses")
+        valid_pose_files = os.listdir(valid_pose_dir)
+        filtered_files  = [file for file in valid_pose_files if file.endswith(".txt")]
+        if len(filtered_files) > 0:
+            for file in filtered_files:
+                os.remove(join(valid_pose_dir, file))
+    else:
+        os.mkdir("valid_poses")
+
+    for i in range(1, objectnumber + 1):
+        with open(join(dirpath, "valid_poses/" + str(i) + ".txt"), 'w') as f:
+            pass
+
+def loadOcclusionData(filepath):
+    # Loads information on occlusions from the given .txt file, if present.
+
+    print("Loading occlusion data..")
+    occlusionDict = {}
+    if os.path.exists(filepath):
+        with open(filepath, 'r') as file:
+            for line in file:
+                (key, val) = line.split()
+                occlusionDict[key] = val
+    else:
+        print("No occlusion data found.")
+    
+    return occlusionDict
+
+def correctMasks(basepath, occlusionData, nameToMaskDict):
+    # Corrects masks using occlusion data.
+
+    print("Correcting masks with occlusion data...")
+
+    for key in occlusionData.keys():
+        imgname = str(key) + ".png"
+        print("Currently correcting: " + imgname, end='\r')
+
+        imagepath = join(basepath, "merged_masks/" + imgname)
+        maskValue = nameToMaskDict[occlusionData[key]]
+
+        image = cv2.imread(imagepath)
+        low = np.array([maskValue - 5, maskValue - 5, maskValue - 5])
+        high = np.array([maskValue + 5, maskValue + 5, maskValue + 5])
+
+        mask = cv2.inRange(image, low, high)
+        image[mask > 0] = (255, 255, 255)
+        cv2.imwrite(imagepath, image)
+
+
 
 class ObjectData:
+    # Class used to simplify storing object data and creating dictionaries.
 
     def __init__(self, ID):
         self.ID = ID
